unix: When specifying a file as argument pass absolute file path to lexer
This fixes an inconvenient inconsistency when passing a relative path as argument:
the argument was passed to the lexer as-is resulting in __file__ being equal to
the original path passed in.
All other occurrences of __file__ (when using import) however use the fully
resolved absolute path.

Also renamed basedir to filepath since that reflects what it is, and renamed p to basedir in turn.