stmhal, uart: Overhaul UART to act like a stream object.
send/recv are converted to write/read, and the stream protocol is used
to implement read, readall, readline, write and iteration.  readchar and
writechar are also added to the UART, as well as readinto.

This unifies the UART with other stream objects, makes it more powerful,
reduces code size, and retains all functionality, including 9-bit
read/writes.