unix: Provide "fast" target to build interpreter for benchmarking.
This build is not intended for real use, but rather for benchmarking,
and may have random features enabled/disabled to get high scores
in synthetic benchmarks. The intent is to show/prove that MicroPython
codebase can compete with CPython, when configured appropriately. But
the main MicroPython aim still remains to optimize for memory usage
(which inevitibly leads to performance degradation in some areas on
some workloads).